+++ modules/core/src/main/java/io/github/palexdev/mfxcore/enums/SelectionMode.java
@@ -1,0 +1,10 @@
+package io.github.palexdev.mfxcore.enums;
+
+/**
+ * Enumeration to distinguish between the two common selection modes, SINGLE and MULTIPLE.
+ */
+public enum SelectionMode {
+    SINGLE,
+    MULTIPLE,
+    ;
+}